requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Deprecated==1.2.10
Flask==1.1.2
Flask-SocketIO==4.3.1
fusionauth-client==1.19.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2==2.7.7
python-engineio==3.13.2
python-socketio==4.6.0
requests==2.24.0
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
